added redis
