+++ package.json
@@ -19,6 +19,7 @@
     "@radix-ui/react-toast": "^1.2.5",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "ioredis": "^5.4.2",
     "langfuse": "^3.32.0",
     "lucide-react": "^0.460.0",
     "motion": "^11.18.0",
